--- BoEG/Assets/Scripts/Framework/Core/ComponentManip.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 935711156bdb4ebfbbffdac2a1286c9e
-timeCreated: 1610618281
