+++ BoEG/Assets/Scripts/Framework/Core/Trigger/TriggerUtility.cs
@@ -3,13 +3,8 @@ using UnityEngine;
 
 namespace MobaGame.Framework.Core.Trigger
 {
-    public class TriggerHelper<T> where T : Collider
-    {
-        public T Collider { get; set; }
-        public Trigger Trigger { get; set; }
-    }
 
-    public static class TriggerUtility
+	public static class TriggerUtility
     {
         public static Transform GetTriggerContainer(Actor actor)
         {
